cleanup sds tests
