+++ sds/src/scanner/mod.rs
@@ -34,6 +34,9 @@ pub mod regex_rule;
 pub mod scope;
 pub mod shared_pool;
 
+#[cfg(test)]
+mod test;
+
 pub struct StringMatch {
     pub start: usize,
     pub end: usize,
@@ -882,2117 +885,4 @@ fn is_false_positive_match(
         };
     }
     false
-}
-
-#[cfg(test)]
-mod test {
-    use super::CompiledRuleDyn;
-    use super::*;
-    use super::{MatchEmitter, ScannerBuilder, StringMatch};
-    use crate::match_action::{MatchAction, MatchActionValidationError};
-
-    use crate::match_validation::config::{AwsConfig, AwsType, MatchValidationType};
-
-    use crate::match_validation::config::HttpValidatorConfigBuilder;
-    use crate::match_validation::validator_utils::generate_aws_headers_and_body;
-    use crate::observability::labels::Labels;
-    use crate::scanner::regex_rule::config::{
-        ProximityKeywordsConfig, RegexRuleConfig, SecondaryValidator, SecondaryValidator::*,
-    };
-    use crate::scanner::scope::Scope;
-    use crate::scanner::{get_next_regex_start, CreateScannerError, Scanner};
-    use crate::scoped_ruleset::ExclusionCheck;
-    use crate::validation::RegexValidationError;
-    use std::{fmt, time::Duration};
-
-    use crate::{simple_event::SimpleEvent, PartialRedactDirection, Path, PathSegment, RuleMatch};
-    use crate::{Encoding, Utf8Encoding};
-    use ahash::AHashSet;
-    use chrono::Utc;
-    use httpmock::{Method::GET, Method::POST, MockServer};
-
-    use regex_automata::Match;
-    use std::collections::BTreeMap;
-
-    use super::CompiledRule;
-    use super::RuleConfig;
-
-    pub struct DumbRuleConfig {}
-
-    pub struct DumbCompiledRule {
-        pub match_action: MatchAction,
-        pub scope: Scope,
-    }
-
-    impl CompiledRule for DumbCompiledRule {
-        type GroupData = ();
-        type GroupConfig = ();
-        type RuleScanCache = ();
-
-        fn get_match_action(&self) -> &MatchAction {
-            &self.match_action
-        }
-        fn get_scope(&self) -> &Scope {
-            &self.scope
-        }
-
-        fn create_group_data(_: &Labels) {}
-        fn create_group_config() {}
-        fn create_rule_scan_cache() {}
-
-        fn get_string_matches(
-            &self,
-            _content: &str,
-            _path: &Path,
-            _regex_caches: &mut RegexCaches,
-            _group_data: &mut Self::GroupData,
-            _group_config: &Self::GroupConfig,
-            _rule_scan_cache: &mut Self::RuleScanCache,
-            _exclusion_check: &ExclusionCheck<'_>,
-            _excluded_matches: &mut AHashSet<String>,
-            match_emitter: &mut dyn MatchEmitter,
-            _: Option<&Vec<(usize, usize)>>,
-        ) {
-            match_emitter.emit(StringMatch { start: 10, end: 16 });
-        }
-
-        fn get_match_validation_type(&self) -> Option<&MatchValidationType> {
-            None
-        }
-
-        fn get_internal_match_validation_type(&self) -> Option<&InternalMatchValidationType> {
-            None
-        }
-        fn process_scanner_config(&self, _: &mut Self::GroupConfig) {
-            // do nothing
-        }
-    }
-
-    impl RuleConfig for DumbRuleConfig {
-        fn convert_to_compiled_rule(
-            &self,
-            _content: usize,
-            _: Labels,
-        ) -> Result<Box<dyn CompiledRuleDyn>, CreateScannerError> {
-            Ok(Box::new(DumbCompiledRule {
-                match_action: MatchAction::Redact {
-                    replacement: "[REDACTED]".to_string(),
-                },
-                scope: Scope::default(),
-            }))
-        }
-
-        fn get_match_validation_type(&self) -> Option<&MatchValidationType> {
-            None
-        }
-    }
-
-    #[test]
-    fn dumb_custom_rule() {
-        let scanner = ScannerBuilder::new(&[Arc::new(DumbRuleConfig {})])
-            .build()
-            .unwrap();
-
-        let mut input = "this is a secret with random data".to_owned();
-
-        let matched_rules = scanner.scan(&mut input);
-
-        assert_eq!(matched_rules.len(), 1);
-        assert_eq!(input, "this is a [REDACTED] with random data");
-    }
-
-    #[test]
-    fn test_mixed_rules() {
-        let scanner = ScannerBuilder::new(&[
-            Arc::new(DumbRuleConfig {}),
-            RegexRuleConfig::new("secret")
-                .match_action(MatchAction::Redact {
-                    replacement: "[SECRET]".to_string(),
-                })
-                .build(),
-        ])
-        .build()
-        .unwrap();
-
-        let mut input = "this is a dumbss with random data and a secret".to_owned();
-
-        let matched_rules = scanner.scan(&mut input);
-
-        assert_eq!(matched_rules.len(), 2);
-        assert_eq!(
-            input,
-            "this is a [REDACTED] with random data and a [SECRET]"
-        );
-    }
-
-    #[test]
-    fn simple_redaction() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("secret")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .build()])
-        .build()
-        .unwrap();
-
-        let mut input = "text with secret".to_owned();
-
-        let matched_rules = scanner.scan(&mut input);
-
-        assert_eq!(matched_rules.len(), 1);
-        assert_eq!(input, "text with [REDACTED]");
-    }
-
-    #[test]
-    fn simple_redaction_with_additional_labels() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("secret")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .build()])
-        .labels(Labels::new(&[("key".to_string(), "value".to_string())]))
-        .build()
-        .unwrap();
-
-        let mut input = "text with secret".to_owned();
-
-        let matched_rules = scanner.scan(&mut input);
-
-        assert_eq!(matched_rules.len(), 1);
-        assert_eq!(input, "text with [REDACTED]");
-    }
-
-    #[test]
-    fn should_fail_on_compilation_error() {
-        let scanner_result = ScannerBuilder::new(&[RegexRuleConfig::new("\\u").build()]).build();
-        assert!(scanner_result.is_err());
-        assert_eq!(
-            scanner_result.err().unwrap(),
-            CreateScannerError::InvalidRegex(RegexValidationError::InvalidSyntax)
-        )
-    }
-
-    #[test]
-    fn should_validate_zero_char_count_partial_redact() {
-        let scanner_result = ScannerBuilder::new(&[RegexRuleConfig::new("secret")
-            .match_action(MatchAction::PartialRedact {
-                direction: PartialRedactDirection::LastCharacters,
-                character_count: 0,
-            })
-            .build()])
-        .build();
-
-        assert!(scanner_result.is_err());
-        assert_eq!(
-            scanner_result.err().unwrap(),
-            CreateScannerError::InvalidMatchAction(
-                MatchActionValidationError::PartialRedactionNumCharsZero
-            )
-        )
-    }
-
-    #[test]
-    fn multiple_replacements() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("\\d")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .build()])
-        .build()
-        .unwrap();
-
-        let mut content = "testing 1 2 3".to_string();
-
-        let matches = scanner.scan(&mut content);
-
-        assert_eq!(content, "testing [REDACTED] [REDACTED] [REDACTED]");
-        assert_eq!(matches.len(), 3);
-    }
-
-    #[test]
-    fn match_rule_index() {
-        let scanner = ScannerBuilder::new(&[
-            RegexRuleConfig::new("a").build(),
-            RegexRuleConfig::new("b").build(),
-        ])
-        .build()
-        .unwrap();
-
-        let mut content = "a b".to_string();
-
-        let matches = scanner.scan(&mut content);
-
-        assert_eq!(content, "a b");
-        assert_eq!(matches.len(), 2);
-        assert_eq!(matches[0].rule_index, 0);
-        assert_eq!(
-            (
-                matches[0].start_index,
-                matches[0].end_index_exclusive,
-                matches[0].shift_offset
-            ),
-            (0, 1, 0)
-        );
-        assert_eq!(matches[1].rule_index, 1);
-        assert_eq!(
-            (
-                matches[1].start_index,
-                matches[1].end_index_exclusive,
-                matches[1].shift_offset
-            ),
-            (2, 3, 0)
-        );
-    }
-
-    #[test]
-    fn test_indices() {
-        let test_builder = RegexRuleConfig::new("test");
-        let detect_test_rule = test_builder.build();
-        let redact_test_rule = test_builder
-            .match_action(MatchAction::Redact {
-                replacement: "[test]".to_string(),
-            })
-            .build();
-        let redact_test_rule_2 = RegexRuleConfig::new("ab")
-            .match_action(MatchAction::Redact {
-                replacement: "[ab]".to_string(),
-            })
-            .build();
-
-        let test_cases = vec![
-            (vec![detect_test_rule.clone()], "test1", vec![(0, 4, 0)]),
-            (vec![redact_test_rule.clone()], "test2", vec![(0, 6, 2)]),
-            (vec![redact_test_rule.clone()], "xtestx", vec![(1, 7, 2)]),
-            (
-                vec![redact_test_rule.clone()],
-                "xtestxtestx",
-                vec![(1, 7, 2), (8, 14, 4)],
-            ),
-            (
-                vec![redact_test_rule_2.clone()],
-                "xtestxabx",
-                vec![(6, 10, 2)],
-            ),
-            (
-                vec![redact_test_rule_2.clone(), redact_test_rule.clone()],
-                "xtestxabx",
-                vec![(1, 7, 2), (8, 12, 4)],
-            ),
-            (
-                vec![detect_test_rule.clone(), redact_test_rule_2.clone()],
-                "ab-test",
-                vec![(0, 4, 2), (5, 9, 2)],
-            ),
-        ];
-
-        for (rule_config, input, expected_indices) in test_cases {
-            let scanner = ScannerBuilder::new(rule_config.leak()).build().unwrap();
-            let mut input = input.to_string();
-            let matches = scanner.scan(&mut input);
-
-            assert_eq!(matches.len(), expected_indices.len());
-            for (rule_match, expected_range) in matches.iter().zip(expected_indices) {
-                assert_eq!(
-                    (
-                        rule_match.start_index,
-                        rule_match.end_index_exclusive,
-                        rule_match.shift_offset
-                    ),
-                    expected_range
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_included_keywords_match_content() {
-        let redact_test_rule = RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["hello".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[redact_test_rule]).build().unwrap();
-        let mut content = "hello world".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(content, "hello [REDACTED]");
-        assert_eq!(matches.len(), 1);
-
-        let mut content = "he**o world".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(content, "he**o world");
-        assert_eq!(matches.len(), 0);
-
-        let mut content = "world hello world".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(content, "world hello [REDACTED]");
-        assert_eq!(matches.len(), 1);
-    }
-
-    fn build_test_scanner() -> Scanner {
-        let redact_test_rule = RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["awsAccess".to_string(), "access/key".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build();
-
-        return Scanner::builder(&[redact_test_rule]).build().unwrap();
-    }
-
-    #[test]
-    fn test_included_keywords_match_path() {
-        let scanner = build_test_scanner();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([(
-            "aws".to_string(),
-            SimpleEvent::Map(BTreeMap::from([(
-                "access".to_string(),
-                SimpleEvent::String("hello world".to_string()),
-            )])),
-        )]));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_included_keywords_match_path_case_insensitive() {
-        let scanner = build_test_scanner();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([(
-            "access".to_string(),
-            SimpleEvent::Map(BTreeMap::from([(
-                "KEY".to_string(),
-                SimpleEvent::String("hello world".to_string()),
-            )])),
-        )]));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_included_keywords_match_path_camel_case() {
-        let scanner = build_test_scanner();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([(
-            "accessKEY".to_string(),
-            SimpleEvent::String("hello world".to_string()),
-        )]));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([(
-            "AccessKey".to_string(),
-            SimpleEvent::String("hello world".to_string()),
-        )]));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_included_keywords_path_not_matching() {
-        let scanner = build_test_scanner();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([(
-            "aws".to_string(),
-            SimpleEvent::List(vec![
-                SimpleEvent::Map(BTreeMap::from([(
-                    "key".to_string(),
-                    SimpleEvent::String("hello world".to_string()),
-                )])),
-                SimpleEvent::Map(BTreeMap::from([(
-                    "access".to_string(),
-                    SimpleEvent::String("hello".to_string()),
-                )])),
-            ]),
-        )]));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 0);
-    }
-
-    #[test]
-    fn test_included_keywords_path_with_uncaught_separator_symbol() {
-        let scanner = build_test_scanner();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([(
-            "aws%access".to_string(),
-            SimpleEvent::String("hello".to_string()),
-        )]));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 0);
-    }
-
-    #[test]
-    fn test_included_keywords_path_deep() {
-        let scanner = build_test_scanner();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([(
-            "aws".to_string(),
-            SimpleEvent::List(vec![
-                SimpleEvent::Map(BTreeMap::from([(
-                    "key".to_string(),
-                    SimpleEvent::String("hello world".to_string()),
-                )])),
-                SimpleEvent::Map(BTreeMap::from([
-                    (
-                        "access".to_string(),
-                        SimpleEvent::Map(BTreeMap::from([(
-                            "random_key".to_string(),
-                            SimpleEvent::String("hello world".to_string()),
-                        )])),
-                    ),
-                    (
-                        "another_key".to_string(),
-                        SimpleEvent::String("hello world".to_string()),
-                    ),
-                ])),
-            ]),
-        )]));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_blocked_rules() {
-        let redact_test_rule = RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[redact_test_rule]).build().unwrap();
-        let mut content = "hello world".to_string();
-
-        // Scan with no blocked rules
-        let matches = scanner.scan(&mut content);
-        assert_eq!(content, "hello [REDACTED]");
-        assert_eq!(matches.len(), 1);
-
-        // Scan with blocked rules
-        let mut content = "hello world".to_string();
-        let matches = scanner.scan_with_options(
-            &mut content,
-            ScanOptionBuilder::new()
-                .with_blocked_rules_idx(vec![0])
-                .build(),
-        );
-        assert_eq!(content, "hello world");
-        assert_eq!(matches.len(), 0);
-    }
-
-    #[test]
-    fn test_excluded_keywords() {
-        let redact_test_rule = RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec![],
-                excluded_keywords: vec!["hello".to_string()],
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[redact_test_rule]).build().unwrap();
-        let mut content = "hello world".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(content, "hello world");
-        assert_eq!(matches.len(), 0);
-
-        let mut content = "he**o world".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(content, "he**o [REDACTED]");
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_luhn_checksum() {
-        let rule = RegexRuleConfig::new("(\\d{16})|((\\d{4} ){3}\\d{4})").match_action(
-            MatchAction::Redact {
-                replacement: "[credit card]".to_string(),
-            },
-        );
-
-        let rule_with_checksum = rule.validator(SecondaryValidator::LuhnChecksum).build();
-
-        let scanner = ScannerBuilder::new(&[rule.build()]).build().unwrap();
-        let mut content = "4556997807150071  4111 1111 1111 1111".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 2);
-        assert_eq!(content, "[credit card]  [credit card]");
-
-        let scanner = ScannerBuilder::new(&[rule_with_checksum]).build().unwrap();
-        let mut content = "4556997807150071  4111 1111 1111 1111".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "4556997807150071  [credit card]");
-    }
-
-    #[test]
-    fn test_chinese_id_checksum() {
-        let rule = RegexRuleConfig::new("\\d+").match_action(MatchAction::Redact {
-            replacement: "[IDCARD]".to_string(),
-        });
-
-        let rule_with_checksum = rule.validator(ChineseIdChecksum).build();
-
-        let scanner = ScannerBuilder::new(&[rule.build()]).build().unwrap();
-        let mut content = "513231200012121657 513231200012121651".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 2);
-        assert_eq!(content, "[IDCARD] [IDCARD]");
-
-        let scanner = ScannerBuilder::new(&[rule_with_checksum]).build().unwrap();
-        let mut content = "513231200012121657 513231200012121651".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "[IDCARD] 513231200012121651");
-    }
-
-    #[test]
-    fn test_iban_checksum() {
-        let rule_with_checksum = RegexRuleConfig::new("DE[0-9]+")
-            .match_action(MatchAction::Redact {
-                replacement: "[IBAN]".to_string(),
-            })
-            .validator(IbanChecker)
-            .build();
-
-        // Valid content with checksum
-        let mut content = "number=DE44500105175407324931".to_string();
-        let scanner = ScannerBuilder::new(&[rule_with_checksum.clone()])
-            .build()
-            .unwrap();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "number=[IBAN]");
-
-        // Invalid content with checksum
-        let mut content = "number=DE34500105175407324931".to_string();
-        let scanner = ScannerBuilder::new(&[rule_with_checksum.clone()])
-            .build()
-            .unwrap();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 0);
-        assert_eq!(content, "number=DE34500105175407324931");
-    }
-
-    #[test]
-    fn test_github_token_checksum() {
-        let rule = RegexRuleConfig::new("[^ ]+").match_action(MatchAction::Redact {
-            replacement: "[GITHUB]".to_string(),
-        });
-
-        let rule_with_checksum = rule.validator(GithubTokenChecksum).build();
-
-        let scanner = ScannerBuilder::new(&[rule.build()]).build().unwrap();
-        let mut content =
-            "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 ghp_M7H4jxUDDWHP4kZ6A4dxlQYsQIWJuq11T4V5"
-                .to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 2);
-        assert_eq!(content, "[GITHUB] [GITHUB]");
-
-        let scanner = ScannerBuilder::new(&[rule_with_checksum]).build().unwrap();
-        let mut content =
-            "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 ghp_M7H4jxUDDWHP4kZ6A4dxlQYsQIWJuq11T4V5"
-                .to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "[GITHUB] ghp_M7H4jxUDDWHP4kZ6A4dxlQYsQIWJuq11T4V5");
-    }
-
-    #[test]
-    fn test_jwt_expiration_checker() {
-        use crate::secondary_validation::generate_jwt;
-        let rule = RegexRuleConfig::new("[A-Za-z0-9._-]+")
-            .match_action(MatchAction::Redact {
-                replacement: "[JWT]".to_string(),
-            })
-            .validator(JwtExpirationChecker)
-            .build();
-        let scanner = ScannerBuilder::new(&[rule]).build().unwrap();
-        let future_time_as_string = (Utc::now().timestamp() + 1000000).to_string();
-
-        let mut content = generate_jwt(future_time_as_string).to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "[JWT]");
-
-        let past_time_as_string = (Utc::now().timestamp() - 1000000).to_string();
-        let mut content = generate_jwt(past_time_as_string).to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 0);
-    }
-
-    #[test]
-    fn test_nhs_checksum() {
-        let rule_with_checksum = RegexRuleConfig::new(".+")
-            .match_action(MatchAction::Redact {
-                replacement: "[NHS]".to_string(),
-            })
-            .validator(NhsCheckDigit)
-            .build();
-
-        let mut content = "1234567881".to_string();
-        // Test matching NHS number with checksum
-        let scanner = ScannerBuilder::new(&[rule_with_checksum.clone()])
-            .build()
-            .unwrap();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "[NHS]");
-    }
-
-    #[test]
-    fn test_overlapping_mutations() {
-        // This reproduces a bug where overlapping mutations weren't filtered out, resulting in invalid
-        // UTF-8 indices being calculated which resulted in a panic if they were used.
-
-        let rule = RegexRuleConfig::new("hello")
-            .match_action(MatchAction::Redact {
-                replacement: "*".to_string(),
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule.clone(), rule]).build().unwrap();
-        let mut content = "hello world".to_string();
-        let matches = scanner.scan(&mut content);
-        assert_eq!(content, "* world");
-
-        // The rule was cloned, so if this is only 1, the 2nd was filtered out
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_multiple_partial_redactions() {
-        let rule = RegexRuleConfig::new("...")
-            .match_action(MatchAction::PartialRedact {
-                direction: PartialRedactDirection::FirstCharacters,
-                character_count: 1,
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule.clone(), rule]).build().unwrap();
-        let mut content = "hello world".to_string();
-        let matches = scanner.scan(&mut content);
-
-        assert_eq!(matches.len(), 3);
-        assert_eq!(content, "*el*o *orld");
-
-        assert_eq!(
-            matches[0],
-            RuleMatch {
-                rule_index: 0,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::PartialStart,
-                start_index: 0,
-                end_index_exclusive: 3,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-
-        assert_eq!(
-            matches[1],
-            RuleMatch {
-                rule_index: 0,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::PartialStart,
-                start_index: 3,
-                end_index_exclusive: 6,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-
-        assert_eq!(
-            matches[2],
-            RuleMatch {
-                rule_index: 0,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::PartialStart,
-                start_index: 6,
-                end_index_exclusive: 9,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-    }
-
-    #[test]
-    fn assert_scanner_is_sync_send() {
-        // This ensures that the scanner is safe to use from multiple threads.
-        fn assert_send<T: Send + Sync>() {}
-
-        assert_send::<Scanner>();
-    }
-
-    #[test]
-    fn matches_should_take_precedence_over_non_mutating_overlapping_matches() {
-        let rule_0 = RegexRuleConfig::new("...")
-            .match_action(MatchAction::None)
-            .build();
-
-        let rule_1 = RegexRuleConfig::new("...")
-            .match_action(MatchAction::Redact {
-                replacement: "***".to_string(),
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0, rule_1]).build().unwrap();
-        let mut content = "hello world".to_string();
-        let mut matches = scanner.scan(&mut content);
-        matches.sort();
-
-        assert_eq!(matches.len(), 3);
-        assert_eq!(content, "*********ld");
-
-        assert_eq!(
-            matches[0],
-            RuleMatch {
-                rule_index: 1,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::Placeholder,
-                start_index: 0,
-                end_index_exclusive: 3,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-
-        assert_eq!(
-            matches[1],
-            RuleMatch {
-                rule_index: 1,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::Placeholder,
-                start_index: 3,
-                end_index_exclusive: 6,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-
-        assert_eq!(
-            matches[2],
-            RuleMatch {
-                rule_index: 1,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::Placeholder,
-                start_index: 6,
-                end_index_exclusive: 9,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-    }
-
-    #[test]
-    fn test_overlapping_mutation_higher_priority() {
-        // A mutating match is a higher priority even if it starts after a non-mutating match
-
-        let rule_0 = RegexRuleConfig::new("abc")
-            .match_action(MatchAction::None)
-            .build();
-
-        let rule_1 = RegexRuleConfig::new("bcd")
-            .match_action(MatchAction::Redact {
-                replacement: "***".to_string(),
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0, rule_1]).build().unwrap();
-        let mut content = "abcdef".to_string();
-        let mut matches = scanner.scan(&mut content);
-        matches.sort();
-
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "a***ef");
-
-        assert_eq!(
-            matches[0],
-            RuleMatch {
-                rule_index: 1,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::Placeholder,
-                start_index: 1,
-                end_index_exclusive: 4,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-    }
-
-    #[test]
-    fn test_overlapping_start_offset() {
-        // The match that starts first is used (if the mutation is the same)
-
-        let rule_0 = RegexRuleConfig::new("abc")
-            .match_action(MatchAction::None)
-            .build();
-
-        let rule_1 = RegexRuleConfig::new("bcd")
-            .match_action(MatchAction::None)
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0, rule_1]).build().unwrap();
-        let mut content = "abcdef".to_string();
-        let mut matches = scanner.scan(&mut content);
-        matches.sort();
-
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "abcdef");
-
-        assert_eq!(
-            matches[0],
-            RuleMatch {
-                rule_index: 0,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::None,
-                start_index: 0,
-                end_index_exclusive: 3,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-    }
-
-    #[test]
-    fn test_overlapping_length() {
-        // If 2 matches have the same mutation and same start, the longer one is taken
-
-        let rule_0 = RegexRuleConfig::new("abc")
-            .match_action(MatchAction::None)
-            .build();
-
-        let rule_1 = RegexRuleConfig::new("abcd")
-            .match_action(MatchAction::None)
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0, rule_1]).build().unwrap();
-        let mut content = "abcdef".to_string();
-        let mut matches = scanner.scan(&mut content);
-        matches.sort();
-
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "abcdef");
-
-        assert_eq!(
-            matches[0],
-            RuleMatch {
-                rule_index: 1,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::None,
-                start_index: 0,
-                end_index_exclusive: 4,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-    }
-
-    #[test]
-    fn test_overlapping_rule_order() {
-        // If 2 matches have the same mutation, same start, and the same length, the one with the lower rule index is used
-
-        let rule_0 = RegexRuleConfig::new("abc")
-            .match_action(MatchAction::None)
-            .build();
-
-        let rule_1 = RegexRuleConfig::new("abc")
-            .match_action(MatchAction::None)
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0, rule_1]).build().unwrap();
-        let mut content = "abcdef".to_string();
-        let mut matches = scanner.scan(&mut content);
-        matches.sort();
-
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "abcdef");
-
-        assert_eq!(
-            matches[0],
-            RuleMatch {
-                rule_index: 0,
-                path: Path::root(),
-                replacement_type: crate::ReplacementType::None,
-                start_index: 0,
-                end_index_exclusive: 3,
-                shift_offset: 0,
-
-                match_value: None,
-
-                match_status: MatchStatus::NotAvailable,
-            }
-        );
-    }
-
-    #[test]
-    fn should_skip_match_when_present_in_excluded_matches() {
-        // If 2 matches have the same mutation and same start, the longer one is taken
-
-        let rule_0 = RegexRuleConfig::new("b.*")
-            .scope(Scope::exclude(vec![Path::from(vec![PathSegment::Field(
-                "test".into(),
-            )])]))
-            .match_action(MatchAction::Redact {
-                replacement: "[scrub]".to_string(),
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0]).build().unwrap();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([
-            (
-                "a-match".to_string(),
-                SimpleEvent::String("bcdef".to_string()),
-            ),
-            (
-                "z-match".to_string(),
-                SimpleEvent::String("bcdef".to_string()),
-            ),
-            ("test".to_string(), SimpleEvent::String("bcdef".to_string())),
-        ]));
-
-        let matches = scanner.scan(&mut content);
-
-        // "test" is excluded because it matches the excluded scope.
-        // Both "a-match" and "z-match" are excluded due to having the
-        // same match value as "test" (multi-pass V0)
-        assert_eq!(matches.len(), 0);
-    }
-
-    #[test]
-    fn should_be_able_to_disable_multipass_v0() {
-        let rule_0 = RegexRuleConfig::new("b.*")
-            .scope(Scope::exclude(vec![Path::from(vec![PathSegment::Field(
-                "test".into(),
-            )])]))
-            .match_action(MatchAction::Redact {
-                replacement: "[scrub]".to_string(),
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0])
-            .with_multipass_v0(false)
-            .build()
-            .unwrap();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([
-            (
-                "a-match".to_string(),
-                SimpleEvent::String("bcdef".to_string()),
-            ),
-            (
-                "z-match".to_string(),
-                SimpleEvent::String("bcdef".to_string()),
-            ),
-            ("test".to_string(), SimpleEvent::String("bcdef".to_string())),
-        ]));
-
-        let matches = scanner.scan(&mut content);
-
-        // "test" is excluded because it matches the excluded scope.
-        // Both "a-match" and "z-match" are kept since multipass V0 is disabled
-        assert_eq!(matches.len(), 2);
-    }
-
-    #[test]
-    fn should_not_exclude_false_positive_matches() {
-        // If a match in an excluded scope is a false-positive due to keyword proximity matching,
-        // it is not saved in the excluded matches.
-
-        let rule_0 = RegexRuleConfig::new("b.*")
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["secret".to_string()],
-                excluded_keywords: vec![],
-            })
-            .scope(Scope::exclude(vec![Path::from(vec![PathSegment::Field(
-                "test".into(),
-            )])]))
-            .match_action(MatchAction::Redact {
-                replacement: "[scrub]".to_string(),
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0]).build().unwrap();
-
-        let mut content = SimpleEvent::Map(BTreeMap::from([
-            (
-                "message".to_string(),
-                SimpleEvent::String("secret abcdef".to_string()),
-            ),
-            ("test".to_string(), SimpleEvent::String("bcdef".to_string())),
-        ]));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_calculate_indices_is_called_with_sorted_start_index() {
-        // A custom "Event" implementation is used here to use a different encoding that asserts the indices are in order
-        struct OrderAssertEvent(SimpleEvent);
-
-        impl crate::Event for OrderAssertEvent {
-            type Encoding = AssertOrderEncoding;
-
-            fn visit_event<'a>(&'a mut self, visitor: &mut impl crate::EventVisitor<'a>) {
-                self.0.visit_event(visitor)
-            }
-
-            fn visit_string_mut(&mut self, path: &Path, visit: impl FnMut(&mut String) -> bool) {
-                self.0.visit_string_mut(path, visit)
-            }
-        }
-
-        struct AssertOrderEncoding;
-
-        impl Encoding for AssertOrderEncoding {
-            type Index = <Utf8Encoding as Encoding>::Index;
-            type IndexShift = <Utf8Encoding as Encoding>::IndexShift;
-
-            fn zero_index() -> Self::Index {
-                <Utf8Encoding as Encoding>::zero_index()
-            }
-
-            fn zero_shift() -> Self::IndexShift {
-                <Utf8Encoding as Encoding>::zero_shift()
-            }
-
-            fn get_index(value: &Self::Index, utf8_index: usize) -> usize {
-                <Utf8Encoding as Encoding>::get_index(value, utf8_index)
-            }
-
-            fn calculate_indices<'a>(
-                _content: &str,
-                match_visitor: impl Iterator<Item = crate::EncodeIndices<'a, Self>>,
-            ) {
-                let mut prev_start = 0;
-                for indices in match_visitor {
-                    assert!(
-                        indices.utf8_start >= prev_start,
-                        "Indices are not in order."
-                    );
-                    prev_start = indices.utf8_start;
-                }
-            }
-
-            fn adjust_shift(shift: &mut Self::IndexShift, before: &str, after: &str) {
-                <Utf8Encoding as Encoding>::adjust_shift(shift, before, after)
-            }
-
-            fn get_shift(value: &Self::IndexShift, utf8_shift: isize) -> isize {
-                <Utf8Encoding as Encoding>::get_shift(value, utf8_shift)
-            }
-        }
-
-        // `rule_0` has a match after `rule_1` (out of order)
-        let rule_0 = RegexRuleConfig::new("efg").build();
-        let rule_1 = RegexRuleConfig::new("abc").build();
-
-        let scanner = ScannerBuilder::new(&[rule_0, rule_1]).build().unwrap();
-
-        let mut content = OrderAssertEvent(SimpleEvent::Map(BTreeMap::from([(
-            "message".to_string(),
-            SimpleEvent::String("abc-efg".to_string()),
-        )])));
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 2);
-    }
-
-    #[test]
-    fn test_hash_with_leading_zero() {
-        let rule_0 = RegexRuleConfig::new(".+")
-            .match_action(MatchAction::Hash)
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0]).build().unwrap();
-
-        let mut content =
-            SimpleEvent::String("rand string that has a leading zero after hashing: y".to_string());
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-
-        // normally 09d99e4b6ad0d289, but the leading 0 is removed
-        assert_eq!(content, SimpleEvent::String("9d99e4b6ad0d289".to_string()));
-    }
-
-    #[test]
-    fn test_hash_with_leading_zero_utf16() {
-        #[allow(deprecated)]
-        let rule_0 = RegexRuleConfig::new(".+")
-            .match_action(MatchAction::Utf16Hash)
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0]).build().unwrap();
-
-        let mut content = "rand string that has a leading zero after hashing: S".to_string();
-
-        let matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-
-        // normally 08c3ad1a22e2edb1, but the leading 0 is removed
-        assert_eq!(content, "8c3ad1a22e2edb1");
-    }
-
-    #[test]
-    fn test_internal_overlapping_matches() {
-        // A simple "credit-card rule is modified a bit to allow a multi-char character in the match
-        let rule_0 = RegexRuleConfig::new("([\\d€]+){1}(,\\d+){3}")
-            .match_action(MatchAction::Redact {
-                replacement: "[credit card]".to_string(),
-            })
-            .validator(LuhnChecksum)
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0]).build().unwrap();
-
-        // The first 4 numbers match as a credit-card, but fail the luhn checksum.
-        // The last 4 numbers (which overlap with the first match) pass the checksum.
-        let mut content = "[5€184,5185,5252,5052,5005]".to_string();
-
-        let matches = scanner.scan(&mut content);
-        // This is mostly asserting that the scanner doesn't panic when encountering multibyte characters
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_next_regex_start_after_false_positive() {
-        let content = "          testtest";
-        let regex_match = Match::must(0, 10..14);
-        assert_eq!(get_next_regex_start(content, &regex_match), Some(11));
-    }
-
-    #[test]
-    fn test_excluded_keyword_with_excluded_chars_in_content() {
-        let rule_0 = RegexRuleConfig::new("value")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec![],
-                excluded_keywords: vec!["test".to_string()],
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_0]).build().unwrap();
-
-        // "test" should NOT be detected as an excluded keyword because "-" is ignored, so the word
-        // boundary shouldn't match here
-        let mut content = "x-test=value".to_string();
-
-        let matches = scanner.scan(&mut content);
-        // This should match because "test" is not found, so it's not a false-positive
-        assert_eq!(matches.len(), 1);
-    }
-
-    #[test]
-    fn test_included_keyword_not_match_further_than_look_ahead_character_count() {
-        let redact_test_rule = RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["hello".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[redact_test_rule]).build().unwrap();
-
-        let mut content = "hello [this block is exactly 37 chars long] world".to_string();
-        let matches = scanner.scan(&mut content);
-
-        assert_eq!(matches.len(), 0);
-    }
-
-    #[test]
-    fn test_included_keyword_multiple_matches_in_one_prefix() {
-        let redact_test_rule = RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["hello".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[redact_test_rule]).build().unwrap();
-
-        let mut content = "hello world world".to_string();
-        let matches = scanner.scan(&mut content);
-
-        // Both "world" matches fit within the 30 char prefix.
-        assert_eq!(matches.len(), 2);
-    }
-
-    #[test]
-    fn test_included_keyword_multiple_prefix_matches() {
-        let redact_test_rule = RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["hello".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build();
-
-        let scanner = ScannerBuilder::new(&[redact_test_rule]).build().unwrap();
-
-        let mut content =
-            "hello world [this takes up enough space to separate the prefixes] world hello world"
-                .to_string();
-        let matches = scanner.scan(&mut content);
-
-        // Both "worlds" after a "hello" should match
-        assert_eq!(matches.len(), 2);
-    }
-
-    #[test]
-    fn test_included_keywords_on_start_boundary_with_space_including_word_boundary() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("ab")
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["id".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build()])
-        .build()
-        .unwrap();
-
-        let mut content = "users id   ab".to_string();
-        let matches = scanner.scan(&mut content);
-
-        assert_eq!(matches.len(), 1);
-        assert_eq!(matches[0].start_index, 11);
-        assert_eq!(matches[0].end_index_exclusive, 13);
-    }
-
-    #[test]
-    fn test_included_keywords_on_end_boundary() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("abc")
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["id".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build()])
-        .build()
-        .unwrap();
-
-        let mut content = "users idabc".to_string();
-        let matches = scanner.scan(&mut content);
-
-        assert_eq!(matches.len(), 0);
-    }
-
-    #[test]
-    fn should_not_look_ahead_too_far() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("x")
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 10,
-                included_keywords: vec!["host".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build()])
-        .build()
-        .unwrap();
-
-        let mut content = "host           x".to_string();
-        assert_eq!(scanner.scan(&mut content).len(), 0);
-
-        let mut content = "host      x".to_string();
-        assert_eq!(scanner.scan(&mut content).len(), 1);
-
-        let mut content = "host       x".to_string();
-        assert_eq!(scanner.scan(&mut content).len(), 0);
-
-        let mut content = " host      x".to_string();
-        assert_eq!(scanner.scan(&mut content).len(), 1);
-    }
-
-    #[test]
-    fn should_verify_included_keywords_on_path_even_if_included_keywords_are_in_string() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("world")
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 10,
-                included_keywords: vec!["hello".to_string()],
-                excluded_keywords: vec![],
-            })
-            .build()])
-        .build()
-        .unwrap();
-
-        let mut event = SimpleEvent::Map(BTreeMap::from([(
-            "hello".to_string(),
-            SimpleEvent::String("hello [more than ten characters] world".to_string()),
-        )]));
-
-        // Even though the included keywords are too far from the match in the string
-        // the keyword is present in the path and that should validate the match.
-        assert_eq!(scanner.scan(&mut event).len(), 1);
-    }
-
-    #[test]
-    fn test_included_and_excluded_keyword() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("world")
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 11,
-                included_keywords: vec!["hey".to_string()],
-                excluded_keywords: vec!["hello".to_string()],
-            })
-            .build()])
-        .build()
-        .unwrap();
-
-        // only the included keyword is present
-        let mut content = "hey world".to_string();
-        assert_eq!(scanner.scan(&mut content).len(), 1);
-
-        // only the excluded keyword is present
-        let mut content = "hello world".to_string();
-        assert_eq!(scanner.scan(&mut content).len(), 0);
-
-        // no keyword is present
-        let mut content = "world".to_string();
-        assert_eq!(scanner.scan(&mut content).len(), 0);
-
-        // included and excluded keywords are present
-        let mut content = "hey, hello world".to_string();
-        assert_eq!(scanner.scan(&mut content).len(), 1);
-    }
-
-    #[test]
-    fn test_should_return_match_with_match_validation() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new("http://localhost:8080".to_string())
-                    .build()
-                    .unwrap(),
-            ))
-            .build()])
-        .build()
-        .unwrap();
-
-        let mut content = "hey world".to_string();
-        let rule_match = scanner.scan(&mut content);
-        assert_eq!(rule_match.len(), 1);
-        assert_eq!(content, "hey [REDACTED]");
-        assert_eq!(rule_match[0].match_value, Some("world".to_string()));
-    }
-
-    #[test]
-    fn test_should_error_if_no_match_validation() {
-        let scanner = ScannerBuilder::new(&[RegexRuleConfig::new("world")
-            .match_action(MatchAction::Redact {
-                replacement: "[REDACTED]".to_string(),
-            })
-            .build()])
-        .build()
-        .unwrap();
-
-        let mut content = "hey world".to_string();
-        let mut rule_match = scanner.scan(&mut content);
-        assert_eq!(rule_match.len(), 1);
-        assert_eq!(content, "hey [REDACTED]");
-        assert_eq!(rule_match[0].match_value, None);
-        // Let's call validate and check that it panics
-        let err = scanner.validate_matches(&mut rule_match);
-        assert!(err.is_err());
-    }
-
-    #[test]
-    fn test_should_allocate_match_validator_depending_on_match_type() {
-        use crate::match_validation::config::AwsConfig;
-
-        let rule_aws_id = RegexRuleConfig::new("aws-id")
-            .match_action(MatchAction::Redact {
-                replacement: "[AWS ID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::Aws(AwsType::AwsId))
-            .build();
-        let rule_aws_secret = RegexRuleConfig::new("aws-secret")
-            .match_action(MatchAction::Redact {
-                replacement: "[AWS SECRET]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::Aws(AwsType::AwsSecret(
-                AwsConfig::default(),
-            )))
-            .build();
-
-        let rule_custom_http_1_domain_1 = RegexRuleConfig::new("custom-http1")
-            .match_action(MatchAction::Redact {
-                replacement: "[CUSTOM HTTP1]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new("http://localhost:8080".to_string())
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-
-        let rule_custom_http_2_domain_1 = RegexRuleConfig::new("custom-http2")
-            .match_action(MatchAction::Redact {
-                replacement: "[CUSTOM HTTP2]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new("http://localhost:8080".to_string())
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-
-        let rule_custom_http_domain_2 = RegexRuleConfig::new("custom-http3")
-            .match_action(MatchAction::Redact {
-                replacement: "[CUSTOM HTTP2]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new("http://localhost:8081".to_string())
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-
-        let scanner = ScannerBuilder::new(&[
-            rule_aws_id,
-            rule_aws_secret,
-            rule_custom_http_1_domain_1,
-            rule_custom_http_2_domain_1,
-            rule_custom_http_domain_2,
-        ])
-        .build()
-        .unwrap();
-
-        // Let's check the number of entries in the match validator map
-        let match_validator_map = &scanner.match_validators_per_type;
-        assert_eq!(match_validator_map.len(), 3);
-        // Custom assertion to check if the validators are the same
-        let aws_validator = match_validator_map
-            .get(&InternalMatchValidationType::Aws)
-            .unwrap();
-        let http_2_validator = match_validator_map
-            .get(&InternalMatchValidationType::CustomHttp(vec![
-                "http://localhost:8080".to_string(),
-            ]))
-            .unwrap();
-        let http_1_validator = match_validator_map
-            .get(&InternalMatchValidationType::CustomHttp(vec![
-                "http://localhost:8081".to_string(),
-            ]))
-            .unwrap();
-        assert!(!std::ptr::eq(
-            http_1_validator.as_ref(),
-            http_2_validator.as_ref()
-        ));
-        assert!(!std::ptr::eq(
-            aws_validator.as_ref(),
-            http_2_validator.as_ref()
-        ));
-        assert!(!std::ptr::eq(
-            aws_validator.as_ref(),
-            http_1_validator.as_ref()
-        ));
-    }
-
-    #[test]
-    fn test_aws_id_only_shall_not_validate() {
-        let rule_aws_id = RegexRuleConfig::new("aws_id")
-            .match_action(MatchAction::Redact {
-                replacement: "[AWS_ID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::Aws(AwsType::AwsId))
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_aws_id]).build().unwrap();
-        let mut content = "this is an aws_id".to_string();
-        let mut matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "this is an [AWS_ID]");
-        assert!(scanner.validate_matches(&mut matches).is_err());
-        assert_eq!(matches[0].match_status, MatchStatus::NotChecked);
-    }
-
-    #[test]
-    fn test_mock_same_http_validator_several_matches() {
-        let server = MockServer::start();
-
-        // Create a mock on the server.
-        let mock_service_valid = server.mock(|when, then| {
-            when.method(GET)
-                .path("/")
-                .header("authorization", "Bearer valid_match");
-            then.status(200);
-        });
-        let mock_service_invalid = server.mock(|when, then| {
-            when.method(GET)
-                .path("/")
-                .header("authorization", "Bearer invalid_match");
-            then.status(404).header("content-type", "text/html");
-        });
-        let mock_service_error = server.mock(|when, then| {
-            when.method(GET)
-                .path("/")
-                .header("authorization", "Bearer error_match");
-            then.status(500).header("content-type", "text/html");
-        });
-
-        let rule_valid_match = RegexRuleConfig::new("\\bvalid_match\\b")
-            .match_action(MatchAction::Redact {
-                replacement: "[VALID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new(server.url("/").to_string())
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-
-        let rule_invalid_match = RegexRuleConfig::new("\\binvalid_match\\b")
-            .match_action(MatchAction::Redact {
-                replacement: "[INVALID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new(server.url("/").to_string())
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-
-        let rule_error_match = RegexRuleConfig::new("\\berror_match\\b")
-            .match_action(MatchAction::Redact {
-                replacement: "[ERROR]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new(server.url("/").to_string())
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-        let scanner =
-            ScannerBuilder::new(&[rule_valid_match, rule_invalid_match, rule_error_match])
-                .build()
-                .unwrap();
-
-        let mut content =
-            "this is a content with a valid_match an invalid_match and an error_match".to_string();
-        let mut matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 3);
-        assert_eq!(
-            content,
-            "this is a content with a [VALID] an [INVALID] and an [ERROR]"
-        );
-        assert!(scanner.validate_matches(&mut matches).is_ok());
-        mock_service_valid.assert();
-        mock_service_invalid.assert();
-        mock_service_error.assert();
-        assert_eq!(matches[0].match_status, MatchStatus::Valid);
-        assert_eq!(matches[1].match_status, MatchStatus::Invalid);
-        assert_eq!(
-            matches[2].match_status,
-            MatchStatus::Error("Unexpected HTTP status code 500".to_string())
-        );
-    }
-
-    #[test]
-    fn test_mock_http_timeout() {
-        let server = MockServer::start();
-        let _ = server.mock(|when, then| {
-            when.method(GET)
-                .path("/")
-                .header("authorization", "Bearer valid_match");
-            then.status(200);
-        });
-        let rule_valid_match = RegexRuleConfig::new("\\bvalid_match\\b")
-            .match_action(MatchAction::Redact {
-                replacement: "[VALID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new(server.url("/").to_string())
-                    .set_timeout(Duration::from_micros(0))
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_valid_match]).build().unwrap();
-
-        let mut content = "this is a content with a valid_match".to_string();
-        let mut matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(content, "this is a content with a [VALID]");
-        assert!(scanner.validate_matches(&mut matches).is_ok());
-        // This will be in the form "Error making HTTP request: "
-        match &matches[0].match_status {
-            MatchStatus::Error(val) => {
-                assert!(val.starts_with("Error making HTTP request:"));
-            }
-            _ => assert!(false),
-        }
-    }
-    #[test]
-    fn test_mock_multiple_match_validators() {
-        let server = MockServer::start();
-
-        // Create a mock on the server.
-        let mock_http_service_valid = server.mock(|when, then| {
-            when.method(GET).path("/http-service");
-            then.status(200);
-        });
-        let mock_aws_service_valid = server.mock(|when, then| {
-            when.method(POST).path("/aws-service");
-            then.status(200);
-        });
-
-        let rule_valid_match = RegexRuleConfig::new("\\bvalid_match\\b")
-            .match_action(MatchAction::Redact {
-                replacement: "[VALID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new(server.url("/http-service").to_string())
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-
-        let rule_aws_id = RegexRuleConfig::new("\\baws_id\\b")
-            .match_action(MatchAction::Redact {
-                replacement: "[AWS_ID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::Aws(AwsType::AwsId))
-            .build();
-
-        let rule_aws_secret = RegexRuleConfig::new("\\baws_secret\\b")
-            .match_action(MatchAction::Redact {
-                replacement: "[AWS_SECRET]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::Aws(AwsType::AwsSecret(AwsConfig {
-                aws_sts_endpoint: server.url("/aws-service").to_string(),
-                forced_datetime_utc: None,
-                timeout: Duration::from_secs(1),
-            })))
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_valid_match, rule_aws_id, rule_aws_secret])
-            .build()
-            .unwrap();
-
-        let mut content =
-            "this is a content with a valid_match an aws_id and an aws_secret".to_string();
-        let mut matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 3);
-        assert_eq!(
-            content,
-            "this is a content with a [VALID] an [AWS_ID] and an [AWS_SECRET]"
-        );
-        assert!(scanner.validate_matches(&mut matches).is_ok());
-        mock_http_service_valid.assert();
-        mock_aws_service_valid.assert();
-        assert_eq!(matches[0].match_status, MatchStatus::Valid);
-        assert_eq!(matches[1].match_status, MatchStatus::Valid);
-        assert_eq!(matches[2].match_status, MatchStatus::Valid);
-    }
-
-    #[test]
-    fn test_mock_endpoint_with_multiple_hosts() {
-        let server = MockServer::start();
-        // Create a mock on the server.
-        let mock_http_service_us = server.mock(|when, then| {
-            when.method(GET).path("/us-service");
-            then.status(200);
-        });
-        let mock_http_service_eu = server.mock(|when, then| {
-            when.method(GET).path("/eu-service");
-            then.status(403);
-        });
-        let rule_valid_match = RegexRuleConfig::new("\\bvalid_match\\b")
-            .match_action(MatchAction::Redact {
-                replacement: "[VALID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::CustomHttp(
-                HttpValidatorConfigBuilder::new(server.url("/$HOST-service").to_string())
-                    .set_hosts(vec!["us".to_string(), "eu".to_string()])
-                    .build()
-                    .unwrap(),
-            ))
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_valid_match]).build().unwrap();
-        let mut content = "this is a content with a valid_match on multiple hosts".to_string();
-        let mut matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 1);
-        assert_eq!(
-            content,
-            "this is a content with a [VALID] on multiple hosts"
-        );
-        assert!(scanner.validate_matches(&mut matches).is_ok());
-        mock_http_service_us.assert();
-        mock_http_service_eu.assert();
-        assert_eq!(matches[0].match_status, MatchStatus::Valid);
-    }
-
-    #[test]
-    fn test_mock_aws_validator() {
-        let server = MockServer::start();
-        let server_url = server.url("/").to_string();
-
-        // Compute signature for valid match
-        let datetime = chrono::Utc::now();
-
-        let aws_id_valid = "AKIAYYB64AB3GAW3WH79";
-        let aws_id_invalid = "AKIAYYB64AB3GAW3WH70";
-        let aws_id_error = "AKIAYYB64AB3GAW3WH71";
-        let aws_secret_1 = "uYd/WrqSWR6m7rkYsjqGnD3QsmO7hQjDFXPQHMVy";
-        let aws_secret_2 = "uYd/WrqSWR6m7rkYsjqGnD3QsmO7hQjDFXPZHMVy";
-
-        let (_, headers_valid) = generate_aws_headers_and_body(
-            &datetime,
-            server_url.as_str(),
-            aws_id_valid,
-            aws_secret_1,
-        );
-        let valid_authorization = headers_valid.get("authorization").unwrap();
-        let (_, headers_invalid) = generate_aws_headers_and_body(
-            &datetime,
-            server_url.as_str(),
-            aws_id_invalid,
-            aws_secret_1,
-        );
-        let invalid_authorization_1 = headers_invalid.get("authorization").unwrap();
-        let (_, headers_invalid) = generate_aws_headers_and_body(
-            &datetime,
-            server_url.as_str(),
-            aws_id_valid,
-            aws_secret_2,
-        );
-        let invalid_authorization_2 = headers_invalid.get("authorization").unwrap();
-        let (_, headers_error) = generate_aws_headers_and_body(
-            &datetime,
-            server_url.as_str(),
-            aws_id_error,
-            aws_secret_1,
-        );
-        let error_authorization_1 = headers_error.get("authorization").unwrap();
-        let (_, headers_error) = generate_aws_headers_and_body(
-            &datetime,
-            server_url.as_str(),
-            aws_id_error,
-            aws_secret_2,
-        );
-        let error_authorization_2 = headers_error.get("authorization").unwrap();
-        // Create a mock on the server.
-        let mock_service_valid = server.mock(|when, then| {
-            when.method(POST)
-                .path("/")
-                .header("authorization", valid_authorization.to_str().unwrap());
-            then.status(200);
-        });
-        let mock_service_invalid_1 = server.mock(|when, then| {
-            when.method(POST)
-                .path("/")
-                .header("authorization", invalid_authorization_1.to_str().unwrap());
-            then.status(403);
-        });
-        let mock_service_invalid_2 = server.mock(|when, then| {
-            when.method(POST)
-                .path("/")
-                .header("authorization", invalid_authorization_2.to_str().unwrap());
-            then.status(403);
-        });
-        let mock_service_error_1 = server.mock(|when, then| {
-            when.method(POST)
-                .path("/")
-                .header("authorization", error_authorization_1.to_str().unwrap());
-            then.status(500);
-        });
-        let mock_service_error_2 = server.mock(|when, then| {
-            when.method(POST)
-                .path("/")
-                .header("authorization", error_authorization_2.to_str().unwrap());
-            then.status(500);
-        });
-        let rule_aws_id = RegexRuleConfig::new("AKIA[0-9A-Z]{16}")
-            .match_action(MatchAction::Redact {
-                replacement: "[AWS_ID]".to_string(),
-            })
-            .match_validation_type(MatchValidationType::Aws(AwsType::AwsId))
-            .build();
-
-        let rule_aws_secret = RegexRuleConfig::new("[A-Za-z0-9/+]{40}")
-            .match_action(MatchAction::Redact {
-                replacement: "[AWS_SECRET]".to_string(),
-            })
-            .proximity_keywords(ProximityKeywordsConfig {
-                look_ahead_character_count: 30,
-                included_keywords: vec!["aws_secret".to_string()],
-                excluded_keywords: vec![],
-            })
-            .match_validation_type(MatchValidationType::Aws(AwsType::AwsSecret(AwsConfig {
-                aws_sts_endpoint: server_url.clone(),
-                forced_datetime_utc: Some(datetime),
-                timeout: Duration::from_secs(5),
-            })))
-            .build();
-
-        let scanner = ScannerBuilder::new(&[rule_aws_id, rule_aws_secret])
-            .build()
-            .unwrap();
-
-        let mut content = fmt::format(format_args!(
-                                "content with a valid aws_id {}, an invalid aws_id {}, an error aws_id {} and an aws_secret {} and an other aws_secret {}", aws_id_valid, aws_id_invalid, aws_id_error, aws_secret_1, aws_secret_2));
-        let mut matches = scanner.scan(&mut content);
-        assert_eq!(matches.len(), 5);
-        assert_eq!(
-            content,
-            "content with a valid aws_id [AWS_ID], an invalid aws_id [AWS_ID], an error aws_id [AWS_ID] and an aws_secret [AWS_SECRET] and an other aws_secret [AWS_SECRET]"
-        );
-        assert!(scanner.validate_matches(&mut matches).is_ok());
-        mock_service_valid.assert();
-        mock_service_invalid_1.assert();
-        mock_service_invalid_2.assert();
-        mock_service_error_1.assert();
-        mock_service_error_2.assert();
-        assert_eq!(matches[0].match_status, MatchStatus::Valid);
-        assert_eq!(matches[1].match_status, MatchStatus::Invalid);
-        assert_eq!(
-            matches[2].match_status,
-            MatchStatus::Error("Unexpected HTTP status code 500".to_string())
-        );
-        assert_eq!(matches[3].match_status, MatchStatus::Valid);
-        // ID1 + SECRET2 should be in error so it should contain error and not invalid
-        assert_eq!(
-            matches[4].match_status,
-            MatchStatus::Error("Unexpected HTTP status code 500".to_string())
-        );
-    }
-
-    mod metrics_test {
-        use crate::match_action::MatchAction;
-        use crate::scanner::regex_rule::config::{ProximityKeywordsConfig, RegexRuleConfig};
-        use crate::scanner::scope::Scope;
-        use crate::scanner::ScannerBuilder;
-        use crate::{simple_event::SimpleEvent, Path, PathSegment};
-        use metrics::{Key, Label};
-        use metrics_util::debugging::DebugValue;
-        use metrics_util::debugging::DebuggingRecorder;
-        use metrics_util::CompositeKey;
-        use metrics_util::MetricKind::Counter;
-        use std::collections::BTreeMap;
-
-        #[test]
-        fn should_submit_scanning_metrics() {
-            let recorder = DebuggingRecorder::new();
-            let snapshotter = recorder.snapshotter();
-
-            let content_1 = "bcdef";
-            let content_2 = "no match";
-
-            metrics::with_local_recorder(&recorder, || {
-                let rule_0 = RegexRuleConfig::new(content_1)
-                    .match_action(MatchAction::None)
-                    .build();
-
-                let scanner = ScannerBuilder::new(&[rule_0]).build().unwrap();
-                let mut content = SimpleEvent::Map(BTreeMap::from([
-                    (
-                        "key1".to_string(),
-                        SimpleEvent::String(content_1.to_string()),
-                    ),
-                    (
-                        "key2".to_string(),
-                        SimpleEvent::String(content_2.to_string()),
-                    ),
-                ]));
-
-                scanner.scan(&mut content);
-            });
-
-            let snapshot = snapshotter.snapshot().into_hashmap();
-
-            let metric_name = "scanned_events";
-            let metric_value = snapshot
-                .get(&CompositeKey::new(Counter, Key::from_name(metric_name)))
-                .expect("metric not found");
-            assert_eq!(metric_value, &(None, None, DebugValue::Counter(1)));
-
-            let metric_name = "scanning.match_count";
-            let metric_value = snapshot
-                .get(&CompositeKey::new(Counter, Key::from_name(metric_name)))
-                .expect("metric not found");
-            assert_eq!(metric_value, &(None, None, DebugValue::Counter(1)));
-
-            let metric_name = "scanning.duration";
-            let metric_value = snapshot
-                .get(&CompositeKey::new(Counter, Key::from_name(metric_name)))
-                .expect("metric not found");
-            match metric_value.2 {
-                DebugValue::Counter(val) => assert!(val > 0),
-                _ => assert!(false),
-            }
-        }
-
-        #[test]
-        fn should_submit_excluded_match_metric() {
-            let recorder = DebuggingRecorder::new();
-            let snapshotter = recorder.snapshotter();
-
-            metrics::with_local_recorder(&recorder, || {
-                let rule_0 = RegexRuleConfig::new("bcdef")
-                    .scope(Scope::exclude(vec![Path::from(vec![PathSegment::Field(
-                        "test".into(),
-                    )])]))
-                    .match_action(MatchAction::None)
-                    .build();
-
-                let scanner = ScannerBuilder::new(&[rule_0]).build().unwrap();
-                let mut content = SimpleEvent::Map(BTreeMap::from([
-                    // z-match is considered as a false positive here
-                    (
-                        "z-match".to_string(),
-                        SimpleEvent::String("bcdef".to_string()),
-                    ),
-                    ("test".to_string(), SimpleEvent::String("bcdef".to_string())),
-                ]));
-
-                scanner.scan(&mut content);
-            });
-
-            let snapshot = snapshotter.snapshot().into_hashmap();
-
-            let metric_name = "false_positive.multipass.excluded_match";
-            let metric_value = snapshot
-                .get(&CompositeKey::new(Counter, Key::from_name(metric_name)))
-                .expect("metric not found");
-
-            assert_eq!(metric_value, &(None, None, DebugValue::Counter(1)));
-        }
-
-        #[test]
-        fn should_submit_excluded_keywords_metric() {
-            let recorder = DebuggingRecorder::new();
-            let snapshotter = recorder.snapshotter();
-
-            metrics::with_local_recorder(&recorder, || {
-                let redact_test_rule = RegexRuleConfig::new("world")
-                    .match_action(MatchAction::Redact {
-                        replacement: "[REDACTED]".to_string(),
-                    })
-                    .proximity_keywords(ProximityKeywordsConfig {
-                        look_ahead_character_count: 30,
-                        included_keywords: vec![],
-                        excluded_keywords: vec!["hello".to_string()],
-                    })
-                    .build();
-
-                let scanner = ScannerBuilder::new(&[redact_test_rule]).build().unwrap();
-                let mut content = SimpleEvent::Map(BTreeMap::from([(
-                    "test".to_string(),
-                    SimpleEvent::String("hello world".to_string()),
-                )]));
-                scanner.scan(&mut content);
-            });
-
-            let snapshot = snapshotter.snapshot().into_hashmap();
-
-            let metric_name = "false_positive.proximity_keywords";
-
-            let labels = vec![Label::new("type", "excluded_keywords")];
-
-            let metric_value = snapshot
-                .get(&CompositeKey::new(
-                    Counter,
-                    Key::from_parts(metric_name, labels),
-                ))
-                .expect("metric not found");
-
-            assert_eq!(metric_value, &(None, None, DebugValue::Counter(1)));
-        }
-
-        #[test]
-        fn test_regex_match_and_included_keyword_same_index() {
-            let email_rule = RegexRuleConfig::new(".+")
-                .match_action(MatchAction::Redact {
-                    replacement: "[REDACTED]".to_string(),
-                })
-                .proximity_keywords(ProximityKeywordsConfig {
-                    look_ahead_character_count: 30,
-                    included_keywords: vec!["email".to_string()],
-                    excluded_keywords: vec![],
-                })
-                .build();
-
-            let scanner = ScannerBuilder::new(&[email_rule])
-                .with_return_matches(true)
-                .build()
-                .unwrap();
-            let mut content = SimpleEvent::Map(BTreeMap::from([(
-                "message".to_string(),
-                SimpleEvent::String("email=firstname.lastname@acme.com&page2".to_string()),
-            )]));
-            let matches = scanner.scan(&mut content);
-            assert_eq!(matches.len(), 1);
-
-            assert_eq!(
-                matches[0].match_value,
-                Some("=firstname.lastname@acme.com&page2".to_string())
-            );
-        }
-    }
 }
